Latest Merge from Develop to Test

+++ src/app/features/workplace/view-workplace/view-workplace.component.ts
@@ -2,11 +2,13 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Establishment } from '@core/model/establishment.model';
 import { ParentPermissions } from '@core/model/my-workplaces.model';
+import { Roles } from '@core/model/roles.enum';
 import { URLStructure } from '@core/model/url.model';
 import { AlertService } from '@core/services/alert.service';
 import { DialogService } from '@core/services/dialog.service';
 import { EstablishmentService } from '@core/services/establishment.service';
 import { UserService } from '@core/services/user.service';
+import { WorkerService } from '@core/services/worker.service';
 import {
   DeleteWorkplaceDialogComponent,
 } from '@features/workplace/delete-workplace-dialog/delete-workplace-dialog.component';
@@ -21,20 +23,28 @@ export class ViewWorkplaceComponent implements OnInit, OnDestroy {
   public workplace: Establishment;
   public summaryReturnUrl: URLStructure;
   public staffPermission = ParentPermissions.WorkplaceAndStaff;
+  public canDelete: boolean;
+  public canViewStaffRecords: boolean;
+  public totalStaffRecords: number;
   private subscriptions: Subscription = new Subscription();
 
   constructor(
+    private route: ActivatedRoute,
+    private router: Router,
     private alertService: AlertService,
     private dialogService: DialogService,
     private establishmentService: EstablishmentService,
-    private route: ActivatedRoute,
-    private router: Router,
-    private userService: UserService
+    private userService: UserService,
+    private workerService: WorkerService
   ) {}
 
   ngOnInit() {
     this.primaryEstablishment = this.establishmentService.primaryWorkplace;
     this.workplace = this.route.snapshot.data.establishment;
+
+    this.subscriptions.add(
+      this.workerService.getTotalStaffRecords(this.workplace.uid).subscribe(total => (this.totalStaffRecords = total))
+    );
 
     this.summaryReturnUrl = {
       url: ['/workplace', this.workplace.uid],
@@ -44,6 +54,9 @@ export class ViewWorkplaceComponent implements OnInit, OnDestroy {
     this.userService.updateReturnUrl({
       url: ['/workplace', this.workplace.uid],
     });
+
+    this.canDelete = this.primaryEstablishment.isParent && this.userService.loggedInUser.role !== Roles.Read;
+    this.canViewStaffRecords = this.userService.loggedInUser.role !== Roles.Read;
   }
 
   public checkPermission(permission: ParentPermissions) {
@@ -52,6 +65,10 @@ export class ViewWorkplaceComponent implements OnInit, OnDestroy {
 
   public onDeleteWorkplace(event: Event): void {
     event.preventDefault();
+    if (!this.canDelete) {
+      return;
+    }
+
     this.dialogService
       .open(DeleteWorkplaceDialogComponent, { workplaceName: this.workplace.name })
       .afterClosed.subscribe(deleteConfirmed => {
@@ -62,6 +79,10 @@ export class ViewWorkplaceComponent implements OnInit, OnDestroy {
   }
 
   private deleteWorkplace(): void {
+    if (!this.canDelete) {
+      return;
+    }
+
     this.subscriptions.add(
       this.establishmentService.deleteWorkplace(this.workplace.uid).subscribe(
         () => {
